requirements.txt
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
enum34==1.1.6
idna==2.8
importlib-metadata==0.23
lxml==4.4.2
more-itertools==7.2.0
namedlist==1.7
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.0.0
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-rerunfailures==8.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
